+++ hoki/csp/tests/test_utils.py
@@ -44,92 +44,10 @@ def test_metallicity_per_bin():
     expected = np.arange(0.5, 100, 1)
     assert np.isclose(out, expected).all(), "Z per bin has failed"
 
-#############################
-#  Test BPASS File Loading  #
-#############################
 
-
-class TestLoadRates(object):
-
-    # Setup files to load
-    data = hoki.load.model_output(
-        f"{data_path}/supernova-bin-imf135_300.zem5.dat")
-
-    # Check if function loads rates
-    @patch("hoki.load.model_output")
-    def test_load_rates(self, mock_model_output):
-        mock_model_output.return_value = self.data
-        x = utils.load_rates(f"{data_path}", "imf135_300"),\
-            "The rates cannot be initialised."
-
-    # Load rates
-    with patch("hoki.load.model_output") as mock_model_output:
-        mock_model_output.return_value = data
-        x = utils.load_rates(f"{data_path}", "imf135_300")
-
-    # Test wrong inputs
-    def test_file_not_present(self):
-        with pytest.raises(AssertionError):
-            _ = utils.load_rates(f"{data_path}", "imf135_300"),\
-                "The file is not present, but the load function runs."
-
-    def test_wrong_imf(self):
-        with pytest.raises(HokiKeyError):
-            _ = utils.load_rates(f"{data_path}", "i"),\
-                "An unsupported IMF is taken as an input."
-
-    # Test output
-    def test_output_shape(self):
-        assert type(self.x) == pd.DataFrame
-        assert (self.x.columns.get_level_values(0).unique() ==
-                np.array(BPASS_EVENT_TYPES)).all(),\
-            "wrong headers read from the file."
-        assert (self.x.columns.get_level_values(1).unique() ==
-                np.array(BPASS_NUM_METALLICITIES)).all(),\
-            "wrong metallicity header"
-
-    def test_output(self):
-        assert np.isclose(self.x.loc[:, ("Ia", 0.00001)],
-                          self.data["Ia"]).all(),\
-            "Models are not loaded correctly."
-
-
-class TestLoadSpectra(object):
-
-    # Initialise model_output DataFrame
-    # This reduces I/O readings
-    data = hoki.load.model_output(
-        f"{data_path}/spectra-bin-imf135_300.z002.dat").loc[:, slice("6.0", "11.0")]
-
-    # Patch the model_output function
-    @patch("hoki.data_compilers.model_output")
-    def test_compile_spectra(self, mock_model_output):
-
-        # Set the model_output to the DataFrame
-        mock_model_output.return_value = self.data
-
-        spec = utils.load_spectra(f"{data_path}", "imf135_300")
-
-        # Check if compiled file is created
-        assert os.path.isfile(f"{data_path}/all_spectra-bin-imf135_300.npy"),\
-            "No compiled file is created."
-
-        # Check output numpy array
-        npt.assert_allclose(spec[3], self.data.T,
-                            err_msg="Loading of files has failed.")
-
-    def test_load_pickled_file(self):
-
-        spec = utils.load_spectra(f"{data_path}", "imf135_300")
-
-        # Check output numpy array
-        npt.assert_allclose(spec[3], self.data.T,
-                            err_msg="Loading of compiled file has failed.")
-
-        os.remove(f"{data_path}/all_spectra-bin-imf135_300.npy")
 
 ################################
-#  Test Normasise BPASS Files  #
+#  Test Normalise BPASS Files  #
 ################################
