+++ hoki/csp/utils.py
@@ -10,7 +10,6 @@ import pandas as pd
 
 import hoki.load
 from hoki.constants import *
-from hoki.data_compilers import SpectraCompiler
 from hoki.utils.exceptions import *
 
 ########################
@@ -74,138 +73,6 @@ def metallicity_per_bin(Z_function, time_edges):
     Z_values = vec_func(time_edges)
     Z_average = (Z_values[1:] + Z_values[:-1]) / 2
     return np.array(Z_average)
-
-
-########################
-#  BPASS File Loading  #
-########################
-
-
-def load_rates(data_path, imf, binary=True):
-    """
-    Loads the BPASS supernova event files.
-
-    Notes
-    -----
-    The rates are just read from file and not normalised.
-
-    Input
-    -----
-    data_path : `str`
-        The filepath to the folder containing the BPASS data
-    binary : `bool`
-        Use the binary files or just the single stars. Default=True
-    imf : `str`
-        BPASS Identifier of the IMF to be used.
-
-    Returns
-    -------
-    `pandas.DataFrame`
-        A pandas MultiIndex dataframe containing the BPASS number of events
-        per metallicity per type.
-        Usage: rates.loc[log_age, (type, metallicity)]
-
-
-        Note
-        -----
-        This dataframe has the following structure.
-        The index is the log_age as a float.
-        The column is a `pandas.MultiIndex` with the event types
-        (level=0, `float`) and the metallicity (level=1, `float`)
-
-        |Event Type | Ia      | IIP      |  ... | PISNe | low_mass |
-        |Metallicity| 0.00001 | 0.00001  |  ... |  0.04 |    0.04  |
-        | log_age   |---------|----------|------|-------|----------|
-        |    6.0    |
-        |    ...    |                  Event Rate values
-        |    11.0   |
-
-    """
-
-    # This is repeated several times. Make a function?
-    if binary:
-        star = "bin"
-    else:
-        star = "sin"
-
-    # Check IMF
-    if imf not in BPASS_IMFS:
-        raise HokiKeyError(
-            f"{imf} is not a BPASS IMF. Please select a correct IMF.")
-
-    # Create output DataFrame
-    arrays = [BPASS_NUM_METALLICITIES, BPASS_EVENT_TYPES]
-    columns = pd.MultiIndex.from_product(
-        arrays, names=["Metallicicty", "Event Type"])
-    rates = pd.DataFrame(index=np.linspace(6, 11, 51),
-                         columns=columns, dtype=np.float64)
-    rates.index.name = "log_age"
-
-    # load supernova count files
-    for num, metallicity in enumerate(BPASS_METALLICITIES):
-        data = hoki.load.model_output(
-            f"{data_path}/supernova-{star}-{imf}.z{metallicity}.dat")
-        data = data.loc[:, slice(BPASS_EVENT_TYPES[0], BPASS_EVENT_TYPES[-1])]
-        rates.loc[:, (BPASS_NUM_METALLICITIES[num],
-                      slice(None))] = data.to_numpy()
-
-    return rates.swaplevel(0, 1, axis=1)
-
-
-def load_spectra(data_path, imf, binary=True):
-    """
-    Load all BPASS spectra from files.
-
-    Notes
-    -----
-    The first time this function is ran on a folder it will generate a pickle
-    file containing all the BPASS spectra per metallicity for faster loading
-    in the future. It stores the file in the same folder with the name:
-    `all_spectra-[bin/sin]-[imf].pkl`
-
-    The spectra are just read from file and not normalised.
-
-    Input
-    -----
-    data_path : `str`
-        The path to the folder containing the BPASS spectra.
-    binary : `bool`
-        Use the binary files or just the single stars. Default=True
-    imf : `str`
-        BPASS Identifier of the IMF to be used.
-
-    Returns
-    -------
-    spectra : `numpy.ndarray` (13, 51, 100000) [(metallicity, log_age, wavelength)]
-        A 3D numpy array containing all the BPASS spectra for a specific imf
-        and binary or single star population.
-        Usage: spectra[1][2][1000]
-                (gives L_\\odot for Z=0.0001 and log_age=6.2 at 999 Angstrom)
-
-    """
-    if binary:
-        star = "bin"
-    else:
-        star = "sin"
-
-    # check IMF key
-    if imf not in BPASS_IMFS:
-        raise HokiKeyError(
-            f"{imf} is not a BPASS IMF. Please select a correct IMF.")
-
-    # Check if compiled spectra are already present in data folder
-    try:
-        print("Trying to load precompiled file.")
-        spectra = np.load(f"{data_path}/all_spectra-{star}-{imf}.npy")
-        print("Done Loading.")
-
-    # Compile the spectra for faster reading next time
-    except FileNotFoundError:
-        print("Failed")
-        print("Data will be compiled")
-        spec = SpectraCompiler(data_path, data_path, imf)
-        spectra = spec.spectra
-    return spectra
 
 
 ###########################
